+++ composeApp/src/commonMain/kotlin/com/vivimusic/de/ui/screens/TogetherScreen.kt
@@ -1,0 +1,10 @@
+package com.vivimusic.de.ui.screens
+
+import androidx.compose.runtime.Composable
+import com.vivimusic.de.resources.*
+import com.vivimusic.de.ui.EmptyState
+
+@Composable
+fun TogetherScreen() {
+    EmptyState(Res.string.coming_soon)
+}